+++ tests/api/test_account.py
@@ -13,6 +13,7 @@ from django.shortcuts import reverse
 from freezegun import freeze_time
 from prices import Money
 
+from saleor.account import events as account_events
 from saleor.account.models import Address, User
 from saleor.checkout import AddressType
 from saleor.graphql.account.mutations import (
@@ -535,7 +536,7 @@ def test_customer_register(user_api_client):
     content = get_graphql_content(response)
     data = content["data"]["customerRegister"]
     assert not data["errors"]
-    assert User.objects.filter(email=email).exists()
+    new_user = User.objects.get(email=email)
 
     response = user_api_client.post_graphql(query, variables)
     content = get_graphql_content(response)
@@ -543,6 +544,10 @@ def test_customer_register(user_api_client):
     assert data["errors"]
     assert data["errors"][0]["field"] == "email"
     assert data["errors"][0]["message"] == ("User with this Email already exists.")
+
+    customer_creation_event = account_events.CustomerEvent.objects.get()
+    assert customer_creation_event.type == account_events.CustomerEvents.ACCOUNT_CREATED
+    assert customer_creation_event.user == new_user
 
 
 @patch("saleor.dashboard.emails.send_set_password_customer_email.delay")
@@ -610,11 +615,11 @@ def test_customer_create(
     content = get_graphql_content(response)
 
     User = get_user_model()
-    customer = User.objects.get(email=email)
+    new_customer = User.objects.get(email=email)
 
     shipping_address, billing_address = (
-        customer.default_shipping_address,
-        customer.default_billing_address,
+        new_customer.default_shipping_address,
+        new_customer.default_billing_address,
     )
     assert shipping_address == address
     assert billing_address == address
@@ -632,11 +637,15 @@ def test_customer_create(
     assert send_set_password_customer_email_mock.call_count == 1
     args, kwargs = send_set_password_customer_email_mock.call_args
     call_pk = args[0]
-    assert call_pk == customer.pk
+    assert call_pk == new_customer.pk
+
+    customer_creation_event = account_events.CustomerEvent.objects.get()
+    assert customer_creation_event.type == account_events.CustomerEvents.ACCOUNT_CREATED
+    assert customer_creation_event.user == new_customer
 
 
 def test_customer_update(
-    staff_api_client, customer_user, address, permission_manage_users
+    staff_api_client, staff_user, customer_user, address, permission_manage_users
 ):
     query = """
     mutation UpdateCustomer(
@@ -679,7 +688,7 @@ def test_customer_update(
     billing_address_pk = customer_user.default_billing_address.pk
     shipping_address_pk = customer_user.default_shipping_address.pk
 
-    id = graphene.Node.to_global_id("User", customer_user.id)
+    user_id = graphene.Node.to_global_id("User", customer_user.id)
     first_name = "new_first_name"
     last_name = "new_last_name"
     note = "Test update note"
@@ -689,7 +698,7 @@ def test_customer_update(
     address_data["streetAddress1"] = new_street_address
 
     variables = {
-        "id": id,
+        "id": user_id,
         "firstName": first_name,
         "lastName": last_name,
         "isActive": False,
@@ -722,6 +731,93 @@ def test_customer_update(
     assert data["user"]["lastName"] == last_name
     assert data["user"]["note"] == note
     assert not data["user"]["isActive"]
+
+    # The name was changed, an event should have been triggered
+    name_changed_event = account_events.CustomerEvent.objects.get()
+    assert name_changed_event.type == account_events.CustomerEvents.NAME_ASSIGNED
+    assert name_changed_event.user.pk == staff_user.pk
+    assert name_changed_event.parameters == {"message": customer.get_full_name()}
+
+
+def test_customer_update_generates_event_when_changing_email(
+    staff_api_client, staff_user, customer_user, address, permission_manage_users
+):
+    query = """
+    mutation UpdateCustomer(
+            $id: ID!, $firstName: String, $lastName: String, $email: String) {
+        customerUpdate(id: $id, input: {
+            firstName: $firstName,
+            lastName: $lastName,
+            email: $email
+        }) {
+            errors {
+                field
+                message
+            }
+        }
+    }
+    """
+
+    user_id = graphene.Node.to_global_id("User", customer_user.id)
+    address_data = convert_dict_keys_to_camel_case(address.as_data())
+
+    new_street_address = "Updated street address"
+    address_data["streetAddress1"] = new_street_address
+
+    variables = {
+        "id": user_id,
+        "firstName": customer_user.first_name,
+        "lastName": customer_user.last_name,
+        "email": "mirumee@example.com",
+    }
+    staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_users]
+    )
+
+    # The email was changed, an event should have been triggered
+    email_changed_event = account_events.CustomerEvent.objects.get()
+    assert email_changed_event.type == account_events.CustomerEvents.EMAIL_ASSIGNED
+    assert email_changed_event.user.pk == staff_user.pk
+    assert email_changed_event.parameters == {"message": "mirumee@example.com"}
+
+
+def test_customer_update_without_any_changes_generates_no_event(
+    staff_api_client, customer_user, address, permission_manage_users
+):
+    query = """
+    mutation UpdateCustomer(
+            $id: ID!, $firstName: String, $lastName: String, $email: String) {
+        customerUpdate(id: $id, input: {
+            firstName: $firstName,
+            lastName: $lastName,
+            email: $email
+        }) {
+            errors {
+                field
+                message
+            }
+        }
+    }
+    """
+
+    user_id = graphene.Node.to_global_id("User", customer_user.id)
+    address_data = convert_dict_keys_to_camel_case(address.as_data())
+
+    new_street_address = "Updated street address"
+    address_data["streetAddress1"] = new_street_address
+
+    variables = {
+        "id": user_id,
+        "firstName": customer_user.first_name,
+        "lastName": customer_user.last_name,
+        "email": customer_user.email,
+    }
+    staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_users]
+    )
+
+    # No event should have been generated
+    assert not account_events.CustomerEvent.objects.exists()
 
 
 UPDATE_LOGGED_CUSTOMER_QUERY = """
@@ -782,7 +878,7 @@ def test_logged_customer_update_anonymus_user(api_client):
 
 
 @patch(
-    "saleor.graphql.account.utils.customer_events.staff_user_deleted_a_customer_event"
+    "saleor.graphql.account.utils.account_events.staff_user_deleted_a_customer_event"
 )
 def test_customer_delete(
     mocked_deletion_event,
@@ -1043,6 +1139,10 @@ def test_set_password(user_api_client, customer_user):
     customer_user.refresh_from_db()
     assert customer_user.check_password(password)
 
+    password_resent_event = account_events.CustomerEvent.objects.get()
+    assert password_resent_event.type == account_events.CustomerEvents.PASSWORD_RESET
+    assert password_resent_event.user == customer_user
+
 
 @patch("saleor.account.emails.send_password_reset_email.delay")
 def test_password_reset_email(
@@ -1070,8 +1170,10 @@ def test_password_reset_email(
     args, kwargs = send_password_reset_mock.call_args
     call_context = args[0]
     call_email = args[1]
+    call_user_pk = args[2]
     assert call_email == email
     assert "token" in call_context
+    assert call_user_pk == customer_user.pk
 
 
 @patch("saleor.account.emails.send_password_reset_email.delay")