+++ tests/api/test_bulk_delete.py
@@ -237,7 +237,7 @@ def test_delete_collections(
 
 
 @patch(
-    "saleor.graphql.account.utils.customer_events.staff_user_deleted_a_customer_event"
+    "saleor.graphql.account.utils.account_events.staff_user_deleted_a_customer_event"
 )
 def test_delete_customers(
     mocked_deletion_event,